Move clientsets to a dedicated folder
