+++ pkg/clients/dclient/client_test.go
