+++ pkg/clients/dclient/discovery_test.go
