+++ pkg/clients/dclient/utils.go
